[metal] Fix metal erroring on an `array_stride` of 0 (#3538)
