+++ wgpu/src/util/mod.rs
@@ -9,7 +9,6 @@ mod encoder;
 mod indirect;
 mod init;
 
-use std::ops::{Add, Rem, Sub};
 use std::sync::Arc;
 use std::{
     borrow::Cow,
@@ -22,6 +21,7 @@ pub use device::{BufferInitDescriptor, DeviceExt};
 pub use encoder::RenderEncoder;
 pub use indirect::*;
 pub use init::*;
+pub use wgt::math::*;
 
 /// Treat the given byte slice as a SPIR-V module.
 ///
@@ -139,33 +139,5 @@ impl std::ops::Deref for DownloadBuffer {
     type Target = [u8];
     fn deref(&self) -> &[u8] {
         self.1.slice()
-    }
-}
-
-///
-/// Aligns a `value` to an `alignment`.
-///
-/// Returns the first number greater than or equal to `value` that is also a
-/// multiple of `alignment`. If `value` is already a multiple of `alignment`,
-/// `value` will be returned.
-///
-/// # Examples
-///
-/// ```
-/// # use wgpu::util::align_to;
-/// assert_eq!(align_to(253, 16), 256);
-/// assert_eq!(align_to(256, 16), 256);
-/// assert_eq!(align_to(0, 16), 0);
-/// ```
-///
-pub fn align_to<T>(value: T, alignment: T) -> T
-where
-    T: Add<Output = T> + Copy + Default + PartialEq<T> + Rem<Output = T> + Sub<Output = T>,
-{
-    let remainder = value % alignment;
-    if remainder == T::default() {
-        value
-    } else {
-        value + alignment - remainder
     }
 }
